SocialService-Refactoring: Refactoring Resource Controllers, Services and Entities.
