+++ src/main/java/com/umssonline/social/controllers/CommentsRestController.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.models.dto.CreateCommentDto;
-import com.umssonline.social.models.dto.UpdateCommentDto;
+import com.umssonline.social.models.dto.comment.CreateCommentDto;
+import com.umssonline.social.models.dto.comment.UpdateCommentDto;
 import com.umssonline.social.models.entity.Comment;
 import org.springframework.http.ResponseEntity;
 
@@ -9,13 +9,13 @@ import java.util.Collection;
 
 public interface CommentsRestController {
 
-    ResponseEntity<Comment> findById(final Long id) throws Exception;
+    ResponseEntity<Comment> findById(final Long id);
 
-    ResponseEntity<Collection<Comment>> findAll() throws Exception;
+    ResponseEntity<Collection<Comment>> findAll();
 
     Collection<Comment> findByProperty(final String property);
 
-    ResponseEntity<Comment> create(final CreateCommentDto commentDto) throws Exception;
+    ResponseEntity<Comment> create(final CreateCommentDto commentDto);
 
     ResponseEntity<Comment> update(final UpdateCommentDto commentDto);
