+++ src/main/java/com/umssonline/social/controllers/MessagesRestController.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.models.dto.CreateMessageDto;
-import com.umssonline.social.models.dto.UpdateMessageDto;
+import com.umssonline.social.models.dto.comment.CreateMessageDto;
+import com.umssonline.social.models.dto.comment.UpdateMessageDto;
 import com.umssonline.social.models.entity.Message;
 import org.springframework.http.ResponseEntity;
