+++ src/main/java/com/umssonline/social/controllers/RatesRestController.java
@@ -1,7 +1,7 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.models.dto.CreateRateDto;
-import com.umssonline.social.models.dto.UpdateRateDto;
+import com.umssonline.social.models.dto.rate.CreateRateDto;
+import com.umssonline.social.models.dto.rate.UpdateRateDto;
 import com.umssonline.social.models.entity.Rate;
 import org.springframework.http.ResponseEntity;
