--- src/main/java/com/umssonline/social/models/dto/CreateRateDto.java
@@ -1,4 +1,0 @@
-package com.umssonline.social.models.dto;
-
-public class CreateRateDto {
-}
