+++ src/main/java/com/umssonline/social/models/dto/comment/CreateCommentDto.java
@@ -1,0 +1,24 @@
+package com.umssonline.social.models.dto.comment;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+public class CreateCommentDto {
+
+    private Long commentedResourceId;
+    private Long createdById;
+
+    @JsonIgnore
+    private Integer messagesAmount = 0;
+    @JsonIgnore
+    private Boolean isDeleted = false;
+    @JsonIgnore
+    private LocalDateTime createdAt = LocalDateTime.now();
+    @JsonIgnore
+    private LocalDateTime updatedAt = LocalDateTime.now();
+}
