+++ src/main/java/com/umssonline/social/models/dto/comment/UpdateMessageDto.java
@@ -1,22 +1,23 @@
-package com.umssonline.social.models.dto;
+package com.umssonline.social.models.dto.comment;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
-public class CreateMessageDto {
+public class UpdateMessageDto {
 
+    @Id
+    @NotNull
+    private Long id;
     private String text;
-    private Long createdById;
+    private Long updatedById;
 
-    @JsonIgnore
-    private Boolean isDeleted = false;
-    @JsonIgnore
-    private LocalDateTime createdAt = LocalDateTime.now();
     @JsonIgnore
     private LocalDateTime updatedAt = LocalDateTime.now();
 }
