+++ src/main/java/com/umssonline/social/models/dto/rate/UpdateRateDto.java
@@ -1,0 +1,4 @@
+package com.umssonline.social.models.dto.rate;
+
+public class UpdateRateDto {
+}
